+++ .secrets.baseline
@@ -21,6 +21,12 @@
       "name": "CloudantDetector"
     },
     {
+      "name": "DiscordBotTokenDetector"
+    },
+    {
+      "name": "GitHubTokenDetector"
+    },
+    {
       "name": "HexHighEntropyString",
       "limit": 3
     },
@@ -45,6 +51,9 @@
     },
     {
       "name": "PrivateKeyDetector"
+    },
+    {
+      "name": "SendGridDetector"
     },
     {
       "name": "SlackDetector"
@@ -137,20 +146,20 @@
         "is_secret": false
       }
     ],
-    "tests/conftest.py": [
+    "tests/integration/conftest.py": [
       {
         "type": "Secret Keyword",
-        "filename": "tests/conftest.py",
+        "filename": "tests/integration/conftest.py",
         "hashed_secret": "d033e22ae348aeb5660fc2140aec35850c4da997",
         "is_verified": false,
         "line_number": 86,
         "is_secret": false
       }
     ],
-    "tests/test_client.py": [
+    "tests/integration/test_client.py": [
       {
         "type": "Hex High Entropy String",
-        "filename": "tests/test_client.py",
+        "filename": "tests/integration/test_client.py",
         "hashed_secret": "5666c088b494f26cd8f63ace013992f5fc391ce0",
         "is_verified": false,
         "line_number": 18,
@@ -158,7 +167,7 @@
       },
       {
         "type": "Hex High Entropy String",
-        "filename": "tests/test_client.py",
+        "filename": "tests/integration/test_client.py",
         "hashed_secret": "15a6d8daad1278efcaadc0d6e3d1dd2d9ebbc262",
         "is_verified": false,
         "line_number": 239,
@@ -166,7 +175,7 @@
       },
       {
         "type": "Hex High Entropy String",
-        "filename": "tests/test_client.py",
+        "filename": "tests/integration/test_client.py",
         "hashed_secret": "1b0d1a618b5c213dd792bbc3aa96ffa6bc370ef3",
         "is_verified": false,
         "line_number": 303,
@@ -174,7 +183,7 @@
       },
       {
         "type": "Hex High Entropy String",
-        "filename": "tests/test_client.py",
+        "filename": "tests/integration/test_client.py",
         "hashed_secret": "1170ace44158ff189902ff44597efef121623353",
         "is_verified": false,
         "line_number": 515,
@@ -182,7 +191,7 @@
       },
       {
         "type": "Hex High Entropy String",
-        "filename": "tests/test_client.py",
+        "filename": "tests/integration/test_client.py",
         "hashed_secret": "5ddd5d7090744de7a2309660d5d48a931d1c8da8",
         "is_verified": false,
         "line_number": 639,
@@ -190,7 +199,7 @@
       },
       {
         "type": "Hex High Entropy String",
-        "filename": "tests/test_client.py",
+        "filename": "tests/integration/test_client.py",
         "hashed_secret": "ff9c79b737b3ea7386618cc9437d3fb0a772182b",
         "is_verified": false,
         "line_number": 861,
@@ -198,7 +207,7 @@
       },
       {
         "type": "Hex High Entropy String",
-        "filename": "tests/test_client.py",
+        "filename": "tests/integration/test_client.py",
         "hashed_secret": "c8176f1e75e62e15dabaa4087fb7194451c8f6d2",
         "is_verified": false,
         "line_number": 864,
@@ -206,7 +215,7 @@
       },
       {
         "type": "Hex High Entropy String",
-        "filename": "tests/test_client.py",
+        "filename": "tests/integration/test_client.py",
         "hashed_secret": "d5198f8eddb1cbeb437899cd99e5ee97ab8531b4",
         "is_verified": false,
         "line_number": 864,
@@ -214,7 +223,7 @@
       },
       {
         "type": "Hex High Entropy String",
-        "filename": "tests/test_client.py",
+        "filename": "tests/integration/test_client.py",
         "hashed_secret": "02dc196562514eaa3e2feac1f441ccf6ad81e09d",
         "is_verified": false,
         "line_number": 868,
@@ -222,7 +231,7 @@
       },
       {
         "type": "Hex High Entropy String",
-        "filename": "tests/test_client.py",
+        "filename": "tests/integration/test_client.py",
         "hashed_secret": "f1cb2d91a95165a2ab909eadd9f7b65f312c7e2d",
         "is_verified": false,
         "line_number": 869,
@@ -230,7 +239,7 @@
       },
       {
         "type": "Hex High Entropy String",
-        "filename": "tests/test_client.py",
+        "filename": "tests/integration/test_client.py",
         "hashed_secret": "58db546de03270b55a4c889a5c5e6296b29fef25",
         "is_verified": false,
         "line_number": 870,
@@ -238,17 +247,17 @@
       },
       {
         "type": "Hex High Entropy String",
-        "filename": "tests/test_client.py",
+        "filename": "tests/integration/test_client.py",
         "hashed_secret": "b6c0bd08fde409c18760f32bef8705191840c402",
         "is_verified": false,
         "line_number": 871,
         "is_secret": false
       }
     ],
-    "tests/test_schema_migration.py": [
+    "tests/integration/test_schema_migration.py": [
       {
         "type": "Hex High Entropy String",
-        "filename": "tests/test_schema_migration.py",
+        "filename": "tests/integration/test_schema_migration.py",
         "hashed_secret": "5666c088b494f26cd8f63ace013992f5fc391ce0",
         "is_verified": false,
         "line_number": 86,
@@ -256,5 +265,5 @@
       }
     ]
   },
-  "generated_at": "2023-09-18T16:04:20Z"
+  "generated_at": "2023-09-18T17:33:12Z"
 }
