+++ tests/integration/test_client.py
@@ -3,11 +3,11 @@ import random
 import uuid
 
 import pytest
-from swagger_client.rest import ApiException
 
 from indexd.blueprint import dist_get_record
 from indexd.index.blueprint import ACCEPTABLE_HASHES
-from tests.util import assert_blank
+from swagger_client.rest import ApiException
+from tests.integration.util import assert_blank
 
 
 def get_doc(baseid=None, version=None, has_metadata=True):
